webapp/modules/auth/requirements.txt
# Modern Authentication Module Dependencies
# ========================================

# Core authentication
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0

# FastAPI and web framework
fastapi>=0.104.0
starlette>=0.27.0

# Database and data validation
pydantic[email]>=2.5.0
sqlite3  # Built-in with Python

# Security
cryptography>=41.0.0

# Logging and utilities
python-multipart>=0.0.6  # For form data handling 